Refactor: createRoomAPI 함수 및 파일 분리
