+++ FE/src/constants/apiConfig.js
@@ -7,11 +7,13 @@ function getApiBaseUrl() {
 function getApiEndpoints() {
 	const BASE_URL = getApiBaseUrl();
 	const USERS_URL = `${BASE_URL}users/`;
+	const ROOMS_URL = `${BASE_URL}rooms/`;
 
 	return {
 		LOGIN_URL: `${USERS_URL}login/`,
 		REGISTRATION_URL: `${USERS_URL}registration/`,
-		LOGOUT_URL: `${USERS_URL}logout/`
+		LOGOUT_URL: `${USERS_URL}logout/`,
+		ROOMS_URL: `${ROOMS_URL}`
 	};
 }
